[KC-24] added changes - renamed controllers - updated routes - added pagination (set to 100 per page) - added resources and custom requests tor validation - added policies - changed croatian country call number (+385)

+++ app/Http/Controllers/Admin_api/CategoriesController.php
@@ -3,26 +3,32 @@
 namespace App\Http\Controllers\Admin_api;
 
 use App\Http\Controllers\Controller;
+use App\Http\Requests\CreateAdminCategoryRequest;
+use App\Http\Requests\UpdateAdminCategoryRequest;
+use App\Http\Resources\CategoryResource;
 use App\Model\Admin\Category;
 use Exception;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
+use Illuminate\Http\Resources\Json\AnonymousResourceCollection;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Log;
 
-class Categories extends Controller
+class CategoriesController extends Controller
 {
     /**
      * Get categories
      *
-     * @return JsonResponse
+     * @return AnonymousResourceCollection
      */
-    public function list(Request $request): JsonResponse
+    public function index(Request $request): AnonymousResourceCollection
     {
+        $this->authorize('viewAny', Category::class, Auth::user());
+
         try {
             $categories = Category::lookForOriginal($request->filter)->where("parent_id", null)
-            ->orderBy('created_at', 'desc')->get()->load(["pages", "subcategories", "user"]);
-            return response()->json(['data' => $categories], 200);
+            ->orderBy('created_at', 'desc')->with(["pages", "subcategories", "user"])->paginate(100);
+            return CategoryResource::collection($categories);
         } catch (Exception $e) {
             Log::error("Failed to get categories. " . $e->getMessage());
             return response()->json(['message' => $e->getMessage()], 400);
@@ -32,14 +38,11 @@ class Categories extends Controller
     /**
      * Add category
      *
-     * @return JsonResponse
+     * @return CategoryResource
      */
-    public function add(Request $request): JsonResponse
+    public function store(CreateAdminCategoryRequest $request): CategoryResource
     {
-        Log::debug($request->user());
-        $request->validate([
-            'title' => 'required|unique:cms_categories',
-        ]);
+        $this->authorize('create', Category::class, Auth::user());
 
         try {
             $category = new Category();
@@ -48,7 +51,7 @@ class Categories extends Controller
             $category->user_id = Auth::user()->id;
             $category->save();
 
-            return response()->json($category, 200);
+            return new CategoryResource($category);
         } catch (Exception $e) {
             Log::error("Failed to add new category. " . $e->getMessage());
             return response()->json(['message' => $e->getMessage()], 400);
@@ -58,13 +61,16 @@ class Categories extends Controller
     /**
      * Get catrgory
      *
-     * @return JsonResponse
+     * @return CategoryResource
      */
-    public function get(Request $request, int $id): JsonResponse
+    public function show(Request $request, int $id): CategoryResource
     {
         try {
             $category = Category::find($id)->load("subcategories");
-            return response()->json(['data' => $category], 200);
+
+            $this->authorize('view', $category, Auth::user());
+
+            return new CategoryResource($category);
         } catch (Exception $e) {
             Log::error("Failed to get categories. " . $e->getMessage());
             return response()->json(['message' => $e->getMessage()], 400);
@@ -74,20 +80,19 @@ class Categories extends Controller
     /**
      * Edit category
      *
-     * @return JsonResponse
+     * @return CategoryResource
      */
-    public function edit(Request $request, int $id): JsonResponse
+    public function update(UpdateAdminCategoryRequest $request, int $id): CategoryResource
     {
-        $request->validate([
-            'title' => 'required|unique:cms_categories,title,' . $id,
-        ]);
-
         try {
             $category = Category::find($id);
+
+            $this->authorize('update', $category, Auth::user());
+
             $category->title = $request->title;
             $category->save();
 
-            return response()->json($category, 200);
+            return new CategoryResource($category->load(["pages", "subcategories", "user"]));
         } catch (Exception $e) {
             Log::error("Failed to edit category. " . $e->getMessage());
             return response()->json(['message' => $e->getMessage()], 400);
@@ -99,10 +104,13 @@ class Categories extends Controller
      *
      * @return JsonResponse
      */
-    public function delete(Request $request, int $id): JsonResponse
+    public function destroy(Request $request, int $id): JsonResponse
     {
         try {
             $category = Category::find($id);
+
+            $this->authorize('delete', $category, Auth::user());
+
             $category->delete();
 
             return response()->json(['message' => 'success'], 200);
